+++ lib/src/references/src/category.dart
@@ -215,6 +215,12 @@ enum ReferenceCategory {
   @JsonValue('ATS_PA_LOAD_COMBOIO')
   atsPaLoadComboio,
 
+  /// !ATS Reference
+  /// [AllTank Systems] Access to send odometer in execute exits
+  /// ATS_SEND_ODOMETER_IN_EXECUTE_EXITS
+  @JsonValue('ATS_SEND_ODOMETER_IN_EXECUTE_EXITS')
+  atsSendOdometerInExecuteExits,
+
   /// !Mappit Reference
   /// Defines the access to the Mappit module as a operator or mappit employee
   /// Layrz API Reference: MAPPIT_OPERATOR
@@ -434,8 +440,7 @@ enum ReferenceCategory {
 
   /// [fromJson] returns the enum value from a string representation.
   static ReferenceCategory fromJson(String json) {
-    final found = _$ReferenceCategoryEnumMap.entries
-        .firstWhereOrNull((e) => e.value == json);
+    final found = _$ReferenceCategoryEnumMap.entries.firstWhereOrNull((e) => e.value == json);
     return found?.key ?? ReferenceCategory.unknown;
   }